refactor(api): rename decorators -> reflection
