+++ packages/api/src/HalloweenAppDevRunner.spec.ts
@@ -1,6 +1,6 @@
 import * as Koa from "koa";
-import { ExportPathScanner } from "./decorators/ExportPathScanner";
 import { HalloweenAppDevRunner } from "./HalloweenAppDevRunner";
+import { ExportPathScanner } from "./reflection/ExportPathScanner";
 
 jest.mock("koa");
 jest.mock("./decorators/ExportPathScanner");
@@ -13,7 +13,7 @@ describe.skip("HalloweenAppDevRunner", function() {
     mockScan.mockResolvedValue([]);
     process.env.PORT = "3000";
     webserver = new Koa();
-    runner = new HalloweenAppDevRunner(webserver);
+    runner = new HalloweenAppDevRunner();
     runner.run();
   });
