+++ packages/api/src/controllers/UnknownController.ts
@@ -1,6 +1,5 @@
-import { Component } from "../decorators/Component";
-import { Route } from "../decorators/Route";
-import { IndexController } from "./IndexController";
+import { Component } from "../reflection/Component";
+import { Route } from "../reflection/Route";
 
 @Component()
 export class UnknownController {
